Always create PostgreSQL tables using TIMESTAMP WITH TIME ZONE.
The Singer spec best practices asks for timestamps with timezone, so there is no reason that we should not include the information if given.

This change should be backwards compatible. It will not drop the column if it doesn't already include a time zone. That will need to be a manual migration.

Tested locally, using our Zendesk tap fork.