Rewrite column_type() to choose the most general type to use in postgres
JIRA ticket: https://pathlighthq.atlassian.net/browse/FUJ-3612

The following field was causing target postgres to choose "jsonb" as the
type, when it should be "string":

```json
{
    ...
    "previous_value": {
        "type": [
            "null",
            "array",
            "string"
        ],
        "items": {
            "type": [
                "null",
                "string"
            ]
        }
    },
    ...
}
```

This causes some string inputs to fail when we try to insert into that
field, which breaks the tap. Instead, we need to choose "character
varying" as the type, because we can cast to JSON if needed using
custom_sql, but still avoid errors when running the tap.